[refactor][#53] DTO 변화로 인한 테스트 변화
