+++ src/test/java/kr/startoff/backend/domain/user/controller/ProfileControllerTest.java
@@ -29,11 +29,10 @@ import org.springframework.web.context.WebApplicationContext;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import kr.startoff.backend.global.config.SecurityConfig;
-import kr.startoff.backend.domain.user.controller.ProfileController;
 import kr.startoff.backend.domain.user.dto.request.profile.BaekjoonIdRequest;
 import kr.startoff.backend.domain.user.dto.request.profile.BlogUrlRequest;
 import kr.startoff.backend.domain.user.dto.request.profile.GithubUrlRequest;
-import kr.startoff.backend.domain.user.dto.request.profile.NicknameAndIntroduceRequest;
+import kr.startoff.backend.domain.user.dto.request.profile.IntroduceRequest;
 import kr.startoff.backend.global.common.dto.CommonResponse;
 import kr.startoff.backend.domain.tag.service.SkillTagService;
 import kr.startoff.backend.domain.user.service.UserService;
@@ -75,7 +74,7 @@ class ProfileControllerTest {
 
 	@Test
 	void updateUserNicknameAndIntroduceTest() throws Exception {
-		given(userService.updateNicknameAndIntroduce(eq(USER_ID), any(NicknameAndIntroduceRequest.class)))
+		given(userService.updateNickname(eq(USER_ID), any(IntroduceRequest.class)))
 			.willReturn(NEW_NICKNAME);
 		Map<String, String> contentMap = new HashMap<>();
 		contentMap.put("nickname", NEW_NICKNAME);
